Add scalac opts to HouseRulesPlugin.
These are from the [Coding Guideline][0].

[0]: http://www.scala-sbt.org/0.13/docs/Coding-Guideline.html

The idea is that these are the options we always want present, and on a
per-project basis we can opt-out of anything thanks to "-=" and "--=".

Not added is "-Xfatal-warnings", for now, as I think it will always need to be
opted-out of at this point in time.